Finish the script, lint and format

+++ src/docker/utils/prune.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# cspell:word localnet
+# cspell:word toplevel
 
 root_dir=$(git rev-parse --show-toplevel)
 docker_dir=$root_dir/src/docker
@@ -7,6 +9,7 @@ sh_utils=$root_dir/src/sh/utils
 source $sh_utils/colors.sh
 
 yes=''
+reset_localnet=''
 
 ################################################################################
 #                                 Display help                                 #
@@ -16,22 +19,28 @@ display_help_option() {
 	local option=$1
 	local description=$2
 
-	indented_option=$(printf "%-16s" "$option")
+	indented_option=$(printf "%-24s" "$option")
 	colored_option=$(highlight_text "$indented_option")
 	printf "  $colored_option %s\n" "$description"
 }
 
 show_help() {
-	echo "Usage: $0 [COMMAND] [OPTION]"
+	echo
+	executable=$(debug_text $0)
+	options=$(very_dim_text "[OPTIONS]")
+	echo "Usage: $executable $options"
+	echo
 	echo 'Prune the `emojicoin-dot-fun` Docker Compose environment.'
+	echo 'This will delete all `emojicoin` container and volume data.'
 	echo
 	echo "Options:"
-    display_help_option \
-        "-y, --yes" \
-        "Skip the confirmation prompt and force restart/delete the localnet."
-    display_help_option \
-        "-f, --force-restart" \
-        "Skip the confirmation prompt and force restart/delete the localnet."
+	display_help_option \
+		"-y, --yes" \
+		"Skip the confirmation prompt and prune."
+	display_help_option \
+		"-r, --reset-localnet" \
+		"Reset the localnet data as well."
+	echo
 	display_help_option \
 		"-h, --help" \
 		"Display this help message."
@@ -44,9 +53,17 @@ show_help() {
 while [[ $# -gt 0 ]]; do
 	case $1 in
 	--yes) yes=true ;;
-    -y) yes=true ;;
-    --force-restart) yes=true ;;
-    -f) yes=true ;;
+	-y) yes=true ;;
+	-r) reset_localnet=true ;;
+	--reset-localnet) reset_localnet=true ;;
+	-h)
+		show_help
+		exit 0
+		;;
+	--help)
+		show_help
+		exit 0
+		;;
 	*)
 		log_error "Unknown parameter passed: $1"
 		show_help
@@ -56,39 +73,54 @@ while [[ $# -gt 0 ]]; do
 	shift
 done
 
-
 ################################################################################
 #                                Prompt the user                               #
 ################################################################################
-prompted=false
+
 if [ -z "$yes" ]; then
-    prompted=true
-    msg='This script will force restart the localnet and delete all `emojicoin`'
-    msg+=" container and volume data."
-    log_warning "$msg"
-    msg="Are you sure you want to continue? (y/n)"
-    warning=$(log_warning "$msg")
+	reset_msg=""
+	if [ -n "$reset_localnet" ]; then
+		reset_msg="force restart the localnet and delete"
+	else
+		reset_msg="delete"
+	fi
+	msg="This will $reset_msg all \`emojicoin\`"
+	msg+=" container and volume data."
+	log_warning "$msg"
+	msg="Are you sure you want to continue? (y/n)"
+	warning=$(log_warning "$msg")
 
-    input='n'
-    read -r -p "$warning " input
+	input='n'
+	read -r -p "$warning " input
 
-    if [[ "$input" == "y" || "$input" == "Y" || "$input" == "yes" ]]; then
-        yes=true
-    fi
+	if [[ $input == "y" || $input == "Y" || $input == "yes" ]]; then
+		yes=true
+	fi
 fi
 
 if [ -z "$yes" ]; then
-    log_info "Exiting..."
-    exit 0
+	log_info "Exiting..."
+	exit 0
 fi
 
-msg="Removing all containers, volumes, and localnet data."
+msg="Pruning things...🗑️"
 log_info "$msg" $'\n'
 
+# Store the original working directory so we can return to it upon any exit.
+original_cwd=$(pwd)
+
+function cleanup() {
+	cd "$original_cwd" || exit 1
+}
+
+# Call `cleanup` on exit.
+trap cleanup EXIT
 
 ################################################################################
 #                                     Prune                                    #
 ################################################################################
+
+cd $docker_dir
 docker compose -f compose.local.yaml down --volumes
 
 postgres="local-testnet-postgres"
@@ -100,4 +132,6 @@ docker rm -f $postgres --volumes 2>/dev/null
 docker rm -f $api 2>/dev/null
 docker volume rm -f $postgres-data
 
-rm -rf $docker_dir/localnet/.aptos/*
+if [ -n "$reset_localnet" ]; then
+	rm -rf $docker_dir/localnet/.aptos/*
+fi
